refactor: move OpenTelemetry tracing setup to a separate module
